tool/requirements.txt
binance-connector
python-dotenv
eth-account==0.9.0
openpyxl==3.1.0